+++ azext_iot/tests/central/test_iot_central_unit.py
@@ -154,7 +154,10 @@ class TestCentralDeviceProvider:
     _device_template = load_json(FileNames.central_device_template_file)
     _edge_template = load_json(FileNames.central_edge_template_file)
     _device_twin = load_json(FileNames.central_device_twin_file)
+    _device_attestation = load_json(FileNames.central_device_attestation_file)
+    _device_properties = load_json(FileNames.central_device_properties_file)
     _edge_modules = load_json(FileNames.central_edge_modules_file)
+    _device_component = load_json(FileNames.central_device_component_file)
 
     @mock.patch("azext_iot.central.services.device_template")
     @mock.patch("azext_iot.central.services.device")
@@ -252,6 +255,23 @@ class TestCentralDeviceProvider:
             todict(EdgeModule(_module)) for _module in self._edge_modules.get("modules")
         ]
         assert parsed_modules == modules
+
+    @mock.patch("azext_iot.central.services.device")
+    def test_should_list_components(self, mock_device_svc):
+        # setup
+        provider = CentralDeviceProvider(
+            cmd=None, app_id=app_id, api_version=ApiVersion.ga_2022_05_31.value
+        )
+        mock_device_svc.list_device_components.return_value = self._device_component
+
+        # act
+        components = [
+            todict(component) for component in provider.list_device_components("someDeviceId").get("value")
+        ]
+
+        # verify
+        assert mock_device_svc.list_device_components.call_count == 1
+        assert components == self._device_component["value"]
 
     @mock.patch("azext_iot.central.services.device_template")
     @mock.patch("azext_iot.central.services.device")
@@ -359,6 +379,26 @@ class TestCentralDeviceProvider:
         twin = provider.get_device_twin("someDeviceId")
         assert twin == self._device_twin
 
+    @mock.patch("azext_iot.central.services.device")
+    def test_should_return_attestation(self, mock_device_svc):
+        provider = CentralDeviceProvider(
+            cmd=None, app_id=app_id, api_version=ApiVersion.ga_2022_05_31.value
+        )
+        mock_device_svc.get_device_attestation.return_value = self._device_attestation
+
+        attestation = provider.get_device_attestation("someDeviceId")
+        assert attestation == self._device_attestation
+
+    @mock.patch("azext_iot.central.services.device")
+    def test_should_return_properties(self, mock_device_svc):
+        provider = CentralDeviceProvider(
+            cmd=None, app_id=app_id, api_version=ApiVersion.ga_2022_05_31.value
+        )
+        mock_device_svc.get_device_properties_or_telemetry_value.return_value = self._device_properties
+
+        properties = provider.get_device_properties("someDeviceId")
+        assert properties == self._device_properties
+
 
 class TestCentralDeviceGroupProvider:
     _device_groups = [
@@ -403,17 +443,17 @@ class TestCentralDeviceGroupProvider:
     def test_should_update_device_group_name(self, mock_device_group_svc):
         # setup
         provider = CentralDeviceGroupProvider(
-            cmd=None, app_id=app_id, api_version=ApiVersion.v1_1_preview.value
+            cmd=None, app_id=app_id, api_version=ApiVersion.ga_2022_05_31.value
         )
         existing = get_object(
-            self._device_group, "DeviceGroup", api_version=ApiVersion.v1_1_preview.value
+            self._device_group, "DeviceGroup", api_version=ApiVersion.ga_2022_05_31.value
         )
         display_name = "NewDeviceGroupName"
         mock_device_group_svc.get_device_group.return_value = existing
         updated_device_group_dict = deepcopy(self._device_group)
         updated_device_group_dict["displayName"] = display_name
         mock_device_group_svc.update_device_group.return_value = get_object(
-            updated_device_group_dict, "DeviceGroup", api_version=ApiVersion.v1_1_preview.value
+            updated_device_group_dict, "DeviceGroup", api_version=ApiVersion.ga_2022_05_31.value
         )
 
         # act